requirements.txt
numpy
scipy
pillow>=4.1.0
tifffile
h5py
